[UPDATE] Update Scintilla to 2.27.
git-svn-id: svn://svn.tuxfamily.org/svnroot/notepadplus/repository/trunk@791 f5eea248-9336-0410-98b8-ebc06183d4e3

+++ scintilla/cocoa/InfoBar.mm
@@ -46,7 +46,7 @@
 //--------------------------------------------------------------------------------------------------
 
 - (void) selectWithFrame: (NSRect) aRect inView: (NSView*) controlView editor: (NSText*) textObj 
-                delegate:(id) anObject start: (int) selStart length: (int) selLength
+                delegate:(id) anObject start: (NSInteger) selStart length: (NSInteger) selLength
 {
 	aRect = [self drawingRectForBounds: aRect];
 	mIsEditingOrSelecting = YES;
